+++ src/components/viewSections/AmendmentInfo/AmendmentInfo.test.js
@@ -1,9 +1,8 @@
 
-import React from 'react';
-
-import { renderWithIntl } from '@folio/stripes-erm-testing';
-import { KeyValue } from '@folio/stripes-testing';
 import { MemoryRouter } from 'react-router-dom';
+
+import { KeyValue, renderWithIntl } from '@folio/stripes-erm-testing';
+
 import translationsProperties from '../../../../test/helpers';
 import amendment from './testResources';
 import AmendmentInfo from './AmendmentInfo';